--- src/scripts/news.js
@@ -1,1 +1,0 @@
-// Sarah --- News Articles
