+++ src/scripts/news/newsEventListeners.js
@@ -1,15 +1,14 @@
 import printerObjects from "./newsDomPrinter.js";
-import apiManager from "./newsApiManager.js"
-// on page load event
+import fetchCalls from "./newsApiManager.js"
 
+const container = document.querySelector("#newsContainer");
+// on page load event
 const eventListeners = {
   onPageLoadEvent: () => {
     const container = document.querySelector("#newsContainer");
-    // page load TIME STAMP?!
-
     container.innerHTML = printerObjects.createNewsInputField();
     // page load
-    apiManager.getAllEvents()
+    fetchCalls.getAllEvents()
       .then((articles) => {
         // console.log(articles);
         articles.forEach((article) => {
@@ -17,6 +16,57 @@ const eventListeners = {
         });
       });
   },
+  saveCardEvent: () => {
+    const titleValue = document.querySelector("#title-input").value;
+    const urlValue = document.querySelector("#article-input").value;
+    const synopsisValue = document.querySelector("#synopsis-input").value;
+    //   console.log(titleValue, urlValue, synopsisValue)
+    const createArticleObject = {
+      title: titleValue,
+      url: urlValue,
+      synopsis: synopsisValue,
+    };
+    fetchCalls
+      .postEvent(createArticleObject)
+      .then(eventListeners.onPageLoadEvent);
+  },
+  deleteEvent: () => {
+    const primaryKey = event.target.id.split("-")[2];
+    fetchCalls.deleteEvent(primaryKey).then(() => {
+      document.querySelector("#newsContainer").innerHTML = "";
+      container.innerHTML = printerObjects.createNewsInputField();
+      fetch("http://localhost:8088/newsArticles")
+        .then(eventListeners.onPageLoadEvent);
+    });
+  },
+  editEvent1: () => {
+    const primaryKey = event.target.id.split("-")[2];
+    fetchCalls.editEvent(primaryKey).then((article) => {
+      document.querySelector(
+        `#news-article-${primaryKey}`
+      ).innerHTML = printerObjects.editForm(article);
+    });
+  },
+  editEventSave: () => {
+    console.log("you clicked the save changes button", event.target.id);
+    // Get values of inputs
+
+    const editTitleValue = document.querySelector("#edit-title-input").value;
+    const editUrlValue = document.querySelector("#edit-article-input").value;
+    const editSynopsisValue = document.querySelector("#edit-synopsis-input")
+      .value;
+
+    // Store those values in an object
+    const articleObjectEdit = {
+      title: editTitleValue,
+      url: editUrlValue,
+      synopsis: editSynopsisValue,
+      id: event.target.id.split("-")[2],
+    };
+    fetchCalls.updateEvent(articleObjectEdit)
+    .then(eventListeners.onPageLoadEvent);
+
+  }
 };
 
 export default eventListeners;
