--- src/style.css
@@ -1,5 +1,0 @@
-/* .card-border {
-  display: flex;
-  flex-direction: flex;
-  border: 1px solid black;
-} */
